Fix for MacOS plus Command Line feature
Fixed a bug on the Mac where backup files would not be saved.
Added ability to pass specification of TiddlyWiki file on command line.
Without a path, this specifies a different file from index.html in the
content directory. With a full file specification, the file may be
anywhere.
Commented out the window size parameters as more sensible defaults are
selected automatically.