Refactor and drag-and-drop fix.
1. Enable plugin on UserAgent flag. 'settings.js' replaces AppJS file in
data/node_modules/appjs/lib. AppJS project agreed to surface UserAgent
setting in API next version, so this replacement is temporary. This
approach fixes drag-and-drop limitation of earlier version.
2. Refactored saveChanges hijack with allowSave function so plugins can
use this to give appropriate savability test. Will attempt to sell
unconditional part of the plugin back to TiddlyWiki.
3. Delegated Url -> Path translation to Node script and used this to
hijack getLocalPath function. Node script can do this translation
without making assumptions. This approach seems better than allowing
TiddlyWiki to screw it up and then fixing it!
4. Use Node 'path' library for all path manipulation rather than string
functions.
5. Use synchronous file save function rather than asynchronous. This
allows the success flag to be returned as intended and should also be
safer as TiddlyWiki does not expect asychronous save.
6. Removed redundant exports into the Window object.
7. QUERY: index.html and empty.html are now identical - is this what you
intended? If so, should one of them be removed from the source
repository and recreated from the other in creation of the distribution
packages?